+++ lib/apathy_drive/systems/limbs.ex
@@ -4,11 +4,11 @@ defmodule Systems.Limbs do
   import Systems.Text
 
   def equipped_items(character) when is_pid(character) do
-    character |> Components.Limbs.value |> equipped_items
+    character |> Components.Limbs.get_items
   end
 
   def equipped_items(limbs) do
-    Map.values(limbs) |> Enum.map(&(&1["items"])) |> List.flatten |> Enum.uniq |> Enum.map(&Items.find_by_id(&1))
+    Map.values(limbs) |> Enum.map(&(&1["items"])) |> List.flatten |> Enum.uniq
   end
 
   def equipped_weapons(monster) do
@@ -27,7 +27,9 @@ defmodule Systems.Limbs do
 
   def get_limb_names(limbs, item) do
     Map.keys(limbs) |> Enum.filter(fn(limb_name) ->
-      limbs[limb_name]["items"] |> Enum.member?(Components.ID.value(item))
+      items = limbs[limb_name]["items"]
+      "#{inspect items} - #{inspect item}"
+      items |> Enum.member?(item)
     end)
   end