+++ test/apathy_drive_test.exs
@@ -1,7 +1,0 @@
-defmodule ApathyDriveTest do
-  use ExUnit.Case
-
-  test "the truth" do
-    assert(true)
-  end
-end