+++ test/limb_test.exs
@@ -4,60 +4,64 @@ defmodule LimbTest do
   setup data do
     {:ok, character } = Entity.init
 
-    case data[:character] do
-      "open-handed" ->
-        Entity.add_component(character,
-                                         Components.Limbs,
-                                         %{
-                                           "left hand"  => %{"type" => "hand", "items" => []},
-                                           "right hand" => %{"type" => "hand", "items" => []}
-                                         })
-     "wielding one-handed" ->
-       Entity.add_component(character,
-                                        Components.Limbs,
-                                        %{
-                                          "left hand"  => %{"type" => "hand", "items" => ["sword"]},
-                                          "right hand" => %{"type" => "hand", "items" => []}
-                                        })
-     "dual-wielding" ->
-       Entity.add_component(character,
-                                        Components.Limbs,
-                                        %{
-                                          "left hand"  => %{"type" => "hand", "items" => ["sword"]},
-                                          "right hand" => %{"type" => "hand", "items" => ["club"]}
-                                        })
-    "wielding two-handed" ->
-       Entity.add_component(character,
-                                        Components.Limbs,
-                                        %{
-                                          "left hand"  => %{"type" => "hand", "items" => ["staff"]},
-                                          "right hand" => %{"type" => "hand", "items" => ["staff"]}
-                                        })
-    end
-
     {:ok, club } = Entity.init
     Entity.add_component(club, Components.WornOn, %{"hand" => 1})
     Entity.add_component(club, Components.Slot, "weapon")
-    Entity.add_component(club, Components.ID, "club")
+    Entity.add_component(club, Components.ID, 1)
+    Entity.add_component(club, Components.Name, "club")
     Items.add(club)
 
     {:ok, sword } = Entity.init
     Entity.add_component(sword, Components.WornOn, %{"hand" => 1})
     Entity.add_component(sword, Components.Slot, "weapon")
-    Entity.add_component(sword, Components.ID, "sword")
+    Entity.add_component(club, Components.ID, 2)
+    Entity.add_component(club, Components.Name, "shortsword")
     Items.add(sword)
 
     {:ok, staff } = Entity.init
     Entity.add_component(staff, Components.WornOn, %{"hand" => 2})
     Entity.add_component(staff, Components.Slot, "weapon")
-    Entity.add_component(staff, Components.ID, "staff")
+    Entity.add_component(club, Components.ID, 3)
+    Entity.add_component(club, Components.Name, "quarterstaff")
     Items.add(staff)
 
     {:ok, gloves } = Entity.init
     Entity.add_component(gloves, Components.WornOn, %{"hand" => 2})
     Entity.add_component(gloves, Components.Slot, "gauntlets")
-    Entity.add_component(gloves, Components.ID, "gloves")
+    Entity.add_component(club, Components.ID, 4)
+    Entity.add_component(club, Components.Name, "padded gloves")
     Items.add(gloves)
+
+    case data[:character] do
+      "open-handed" ->
+        Entity.add_component(character,
+                             Components.Limbs,
+                             %{
+                               "left hand"  => %{"type" => "hand", "items" => []},
+                               "right hand" => %{"type" => "hand", "items" => []}
+                             })
+     "wielding one-handed" ->
+       Entity.add_component(character,
+                            Components.Limbs,
+                            %{
+                              "left hand"  => %{"type" => "hand", "items" => [sword]},
+                              "right hand" => %{"type" => "hand", "items" => []}
+                            })
+     "dual-wielding" ->
+       Entity.add_component(character,
+                            Components.Limbs,
+                            %{
+                              "left hand"  => %{"type" => "hand", "items" => [sword]},
+                              "right hand" => %{"type" => "hand", "items" => [club]}
+                            })
+    "wielding two-handed" ->
+       Entity.add_component(character,
+                            Components.Limbs,
+                            %{
+                              "left hand"  => %{"type" => "hand", "items" => [staff]},
+                              "right hand" => %{"type" => "hand", "items" => [staff]}
+                            })
+    end
 
     { :ok, %{character: character, club: club, sword: sword, staff: staff, gloves: gloves} }
   end
@@ -72,7 +76,7 @@ defmodule LimbTest do
     Components.Limbs.equip(data[:character], data[:club])
 
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["club"]},
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:club]]},
                                                          "right hand" => %{"type" => "hand", "items" => []}
                                                         }
   end
@@ -88,8 +92,8 @@ defmodule LimbTest do
     Components.Limbs.equip(data[:character], data[:sword])
 
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["club"]},
-                                                         "right hand" => %{"type" => "hand", "items" => ["sword"]}
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:club]]},
+                                                         "right hand" => %{"type" => "hand", "items" => [data[:sword]]}
                                                         }
   end
 
@@ -103,97 +107,97 @@ defmodule LimbTest do
     Components.Limbs.equip(data[:character], data[:staff])
 
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["staff"]},
-                                                         "right hand" => %{"type" => "hand", "items" => ["staff"]}
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:staff]]},
+                                                         "right hand" => %{"type" => "hand", "items" => [data[:staff]]}
                                                         }
   end
 
   @tag character: "wielding one-handed"
   test "an entity wielding a one-handed weapon will remove it to wear a two handed weapon", data do
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["sword"]},
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:sword]]},
                                                          "right hand" => %{"type" => "hand", "items" => []}
                                                        }
 
     Components.Limbs.equip(data[:character], data[:staff])
 
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["staff"]},
-                                                         "right hand" => %{"type" => "hand", "items" => ["staff"]}
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:staff]]},
+                                                         "right hand" => %{"type" => "hand", "items" => [data[:staff]]}
                                                         }
   end
 
   @tag character: "dual-wielding"
   test "an entity wielding two one-handed weapons will remove them to wear a two handed weapon", data do
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["sword"]},
-                                                         "right hand" => %{"type" => "hand", "items" => ["club"]}
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:sword]]},
+                                                         "right hand" => %{"type" => "hand", "items" => [data[:club]]}
                                                        }
 
     Components.Limbs.equip(data[:character], data[:staff])
 
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["staff"]},
-                                                         "right hand" => %{"type" => "hand", "items" => ["staff"]}
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:staff]]},
+                                                         "right hand" => %{"type" => "hand", "items" => [data[:staff]]}
                                                         }
   end
 
   @tag character: "wielding one-handed"
   test "an entity wielding a one-handed weapon will not remove it to wear hand armour", data do
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["sword"]},
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:sword]]},
                                                          "right hand" => %{"type" => "hand", "items" => []}
                                                        }
 
     Components.Limbs.equip(data[:character], data[:gloves])
 
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["gloves","sword"]},
-                                                         "right hand" => %{"type" => "hand", "items" => ["gloves"]}
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:gloves],data[:sword]]},
+                                                         "right hand" => %{"type" => "hand", "items" => [data[:gloves]]}
                                                         }
   end
 
   @tag character: "dual-wielding"
   test "an entity wielding two one-handed weapons will not remove them to wear hand armour", data do
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["sword"]},
-                                                         "right hand" => %{"type" => "hand", "items" => ["club"]}
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:sword]]},
+                                                         "right hand" => %{"type" => "hand", "items" => [data[:club]]}
                                                        }
 
     Components.Limbs.equip(data[:character], data[:gloves])
 
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["gloves","sword"]},
-                                                         "right hand" => %{"type" => "hand", "items" => ["gloves","club"]}
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:gloves],data[:sword]]},
+                                                         "right hand" => %{"type" => "hand", "items" => [data[:gloves],data[:club]]}
                                                         }
   end
 
   @tag character: "wielding two-handed"
   test "an entity wielding a two-handed weapon will not remove them to wear hand armour", data do
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["staff"]},
-                                                         "right hand" => %{"type" => "hand", "items" => ["staff"]}
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:staff]]},
+                                                         "right hand" => %{"type" => "hand", "items" => [data[:staff]]}
                                                        }
 
     Components.Limbs.equip(data[:character], data[:gloves])
 
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["gloves","staff"]},
-                                                         "right hand" => %{"type" => "hand", "items" => ["gloves","staff"]}
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:gloves],data[:staff]]},
+                                                         "right hand" => %{"type" => "hand", "items" => [data[:gloves],data[:staff]]}
                                                         }
   end
 
   @tag character: "wielding two-handed"
   test "an entity wielding a two-handed weapon will remove it to wield a one-handed weapon", data do
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["staff"]},
-                                                         "right hand" => %{"type" => "hand", "items" => ["staff"]}
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:staff]]},
+                                                         "right hand" => %{"type" => "hand", "items" => [data[:staff]]}
                                                        }
 
     Components.Limbs.equip(data[:character], data[:sword])
 
     assert Components.Limbs.value(data[:character]) == %{
-                                                         "left hand"  => %{"type" => "hand", "items" => ["sword"]},
+                                                         "left hand"  => %{"type" => "hand", "items" => [data[:sword]]},
                                                          "right hand" => %{"type" => "hand", "items" => []}
                                                         }
   end